requirements.txt
alabaster==0.7.12
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
Automat==20.2.0
Babel==2.8.0
backcall==0.1.0
baidu-aip==2.2.18.0
beautifulsoup4==4.8.2
bleach==3.1.1
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==6.7
colorama==0.4.3
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.2
docutils==0.16
entrypoints==0.3
filelock==3.0.12
Flask==1.1.1
gevent==1.4.0
greenlet==0.4.15
html5lib==1.0.1
hyperlink==19.0.0
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
incremental==17.5.0
ipykernel==5.1.4
ipyparallel==6.2.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
itchat==1.3.10
itsdangerous==1.1.0
jedi==0.15.2
Jinja2==2.11.1
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
lxml==4.5.0
MarkupSafe==1.1.1
mechanize==0.4.5
mistune==0.8.4
mysqlclient==1.4.6
nbconvert==5.6.1
nbformat==5.0.4
nose==1.3.7
notebook==6.0.3
numpy==1.18.1
opencv-python==4.2.0.32
packaging==20.1
pandocfilters==1.4.2
parsel==1.5.2
parso==0.5.2
pickleshare==0.7.5
Pillow==7.0.0
playsound==1.2.2
prometheus-client==0.7.1
prompt-toolkit==3.0.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyDispatcher==2.0.5
Pygments==2.5.2
PyHamcrest==2.0.1
pymongo==3.10.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
pyparsing==2.4.6
pypiwin32==223
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.11.3
PyQt5-sip==4.19.19
pyrsistent==0.15.7
pytesseract==0.3.2
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
queuelib==1.5.0
requests==2.22.0
scrapely==0.14.1
Scrapy==1.8.0
selenium==3.141.0
Send2Trash==1.5.0
service-identity==18.1.0
shiboken2==5.12.0
six==1.13.0
snowballstemmer==2.0.0
soupsieve==2.0
Sphinx==2.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.0
terminado==0.8.3
testpath==0.4.4
tinydb==3.2.2
toml==0.10.0
tornado==6.0.4
traitlets==4.3.3
Twisted==19.10.0
urllib3==1.25.7
virtualenv==20.0.7
virtualenvwrapper-win==1.2.6
w3lib==1.21.0
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
zipp==3.0.0
zope.interface==4.7.1
